requirements.txt
python-dotenv
requests
openai
openai-tools